Reinstated daily_connections saving in random_valve_connections case

+++ nodes/odor_randomizer.py
@@ -331,36 +331,35 @@ if ((have_testonly_params and have_training_params) or
 
 ###############################################################################
 
-# TODO maybe just reset at like 5am or check for experiments in last few hours, 
-# to use the same mappings after midnight, but in the same workday
-daily_connections_filename = '.' + time.strftime('%Y%m%d', time.localtime()) + \
-    '_mappings.p'
-if os.path.isfile(daily_connections_filename):
-    '''
-    c = raw_input('Found saved mappings from today. Load them? ' + \
-        '([y]es/[n]o/[d]elete them.')
-    '''
-    c = 'd'
-    if c.lower() == 'y':
-        rospy.loginfo('loading odors and odor->pin mappings from ' + \
-            daily_connections_filename)
-        with open(daily_connections_filename, 'rb') as f:
-            odors, left_pins, right_pins = pickle.load(f)
-        generate_odor_to_pin_connections = False
+if random_valve_connections:
+    # TODO maybe just reset at like 5am or check for experiments in last few
+    # hours, to use the same mappings after midnight, but in the same workday
+    daily_connections_filename = '.' + time.strftime('%Y%m%d',
+        time.localtime()) + '_mappings.p'
 
-    elif c.lower() == 'n':
-        # TODO TODO will this not overwrite anyway? maybe eliminate either this
-        # or the below option, and make clear that the remaining option will
-        # delete existing temporary mappings
-        generate_odor_to_pin_connections = True
+    if os.path.isfile(daily_connections_filename):
+        c = raw_input('Found saved mappings from today. Load them? ' + \
+            '([y]es/[n]o/[d]elete them.')
+        if c.lower() == 'y':
+            rospy.loginfo('loading odors and odor->pin mappings from ' + \
+                daily_connections_filename)
+            with open(daily_connections_filename, 'rb') as f:
+                odors, left_pins, right_pins = pickle.load(f)
+            generate_odor_to_pin_connections = False
 
-    elif c.lower() == 'd':
-        rospy.logwarn('Deleting ' + daily_connections_filename)
-        os.remove(daily_connections_filename)
-        generate_odor_to_pin_connections = True
+        elif c.lower() == 'n':
+            # TODO TODO will this not overwrite anyway? maybe eliminate either
+            # this or the below option, and make clear that the remaining option
+            # will delete existing temporary mappings
+            generate_odor_to_pin_connections = True
 
-    else:
-        raise ValueError('invalid choice')
+        elif c.lower() == 'd':
+            rospy.logwarn('Deleting ' + daily_connections_filename)
+            os.remove(daily_connections_filename)
+            generate_odor_to_pin_connections = True
+
+        else:
+            raise ValueError('invalid choice')
 
 else:
     generate_odor_to_pin_connections = True
